requirements.txt
faiss-cpu
boto3
langchain-community
langchain
streamlit
pypdf
python-dotenv